requirements.txt
plyara
yara-python
